Make connectrpc.com/conformance go-gettable
Right now, the [conformance] repository is not go-gettable due to a
missing redirect. While this repository may not be generally useful to
go get, it is still useful as it allows running the conformance test
servers via `go run`.

This PR adds a redirect akin to all of the other go-get redirects.

[conformance]: https://github.com/connectrpc/conformance