Add clarity on Swift interceptor concurrency (#266)
Adds note about stream interceptor behavior when the interceptor
performs asynchronous work. This is in response to discussion in [this
issue](https://github.com/connectrpc/connect-swift/issues/342) in
Connect-Swift.

While this is not special behavior of Connect-Swift that differs from
expected behavior in other Connect and gRPC implementations, I felt like
including the note would be helpful regardless.

cc @rebello95

Signed-off-by: Eddie Seay <eddie.seay@cfacorp.com>