Add RFC 005: Python (#285)
I'd like to get a Python implementation officially accepted so we can
focus our efforts on it.

This is a successor to #71 which I think is abandoned.

I have a prototype at https://github.com/firetiger-oss/connect-python/
which I'd like to see moved into the connectrpc organization. The
current state of it is that it has synchronous and asynchronous clients,
and a synchronous server (on WSGI).

It supports unary and streaming requests. It passes conformance tests,
so it correctly handles the low-level details, including header and
trailer metadata, with the configs here:
- sync client:
https://github.com/firetiger-oss/connect-python/blob/main/tests/conformance/sync_config.yaml
- async client:
https://github.com/firetiger-oss/connect-python/blob/main/tests/conformance/async_config.yaml
- sync server:
https://github.com/firetiger-oss/connect-python/blob/main/tests/conformance/sync_server_config.yaml

Unsupported things that I want to support, but haven't gotten around to
yet:
- Idempotent GETs
- Compression in the client
- Async server, probably built on Starlette
- httpx client transport for the async client, instead of just aiohttp

I'm open to working on those before blessing this library, or after,
whatever works. I expect compression to be simple. GETs may be a little
complicated since they involve a totally different encoding scheme to
cram things into the URL. The async server is a very significant amount
of work; I have a branch at [this
PR](https://github.com/firetiger-oss/connect-python/pull/9), but I think
I will abandon it as excessively complex, since it tries to work with
ASGI directly.

---------

Signed-off-by: Spencer Nelson <s@spencerwnelson.com>
Co-authored-by: Paul Sachs <11449728+paul-sachs@users.noreply.github.com>
Co-authored-by: Philip K. Warren <pkwarren@users.noreply.github.com>
Co-authored-by: Nick Snyder <nickdsnyder@gmail.com>